+++ packages/phone-number-privacy/combiner/src/legacy/endpoints/signing/get-threshold-signature.ts
@@ -20,7 +20,7 @@ import Logger from 'bunyan'
 import { Request, Response } from 'firebase-functions'
 import fetch, { Response as FetchResponse } from 'node-fetch'
 import { performance, PerformanceObserver } from 'perf_hooks'
-import { BLSCryptographyClient } from '../../../common/bls/bls-cryptography-client'
+import { BLSCryptographyClient } from '../../../common/crypto-clients/bls-cryptography-client'
 import { getContractKit } from '../../../common/web3/contracts'
 import config, { VERSION } from '../../../config'
